ast_walker: Implement cleanup handlers for opcodes
Some opcodes like T_METHOD_CALL need to clean up after themselves when
their value is not used. But this can only be decides higher up the AST
tree.

The T_EXPRESSIONS node checks if the child node is not an assignment
node, and then calls a clean up handler on the child node if necessary.
This handler then can do what is necessary to clean up (for example,
discard items from the stack.)